requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==4.1.5
django-filter==22.1
django-jazzmin==2.6.0
django-split-settings==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.0.0
jwt==1.3.1
Markdown==3.4.1
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
zipp==3.11.0
